Factor out docusign Django app.
