+++ docusign/management/commands/setdocusignkey.py
@@ -1,7 +1,7 @@
 import sys
 from django.core.management.base import BaseCommand
 
-from hpaction.models import DocusignConfig
+from docusign import core
 
 
 class Command(BaseCommand):
@@ -13,7 +13,7 @@ class Command(BaseCommand):
     def handle(self, *args, **options) -> None:
         stdin = options.get('stdin', sys.stdin)
         private_key = stdin.read()
-        cfg = DocusignConfig.objects.get()
+        cfg = core.get_config()
         cfg.private_key = private_key
         cfg.save()
         self.stdout.write("Saved DocuSign private key to database.\n")
