+++ hpaction/docusign.py
@@ -1,14 +1,12 @@
-from typing import Tuple, Union, Optional
-import urllib.parse
+from typing import Tuple
 import base64
-import docusign_esign as docusign
-from django.core.exceptions import ImproperlyConfigured
-from django.utils import timezone
 from django.conf import settings
 
-from project.util.site_util import absolute_reverse
 from users.models import JustfixUser
-from .models import HPActionDocuments, DocusignConfig
+from docusign.core import docusign_client_user_id
+import docusign_esign as dse
+from .models import HPActionDocuments
+
 
 # The recipient ID for the tenant in the signing flow. This appears to be a
 # number local to a specific signing, rather than a globally unique identifier.
@@ -19,141 +17,8 @@ TENANT_RECIPIENT_ID = '1'
 # identifier.
 HPA_DOCUMENT_ID = '1'
 
-# Number of seconds our JWT lasts.
-JWT_EXPIRATION = 3600
 
-# Settings that are required for DocuSign integration to work properly.
-REQUIRED_SETTINGS = [
-    'DOCUSIGN_ACCOUNT_ID',
-    'DOCUSIGN_INTEGRATION_KEY',
-    'DOCUSIGN_USER_ID',
-]
-
-
-def get_account_base_uri(token: Union[str, docusign.OAuthToken]) -> Optional[str]:
-    '''
-    Validate that the given OAuth token has access to the DocuSign
-    account we need, and return the account's base URI.
-
-    If the token doesn't have access to our DocuSign account, return None.
-    '''
-
-    api_client = create_api_client(get_auth_server_url(), token)
-
-    # https://developers.docusign.com/esign-rest-api/guides/authentication/user-info-endpoints
-    response = api_client.call_api("/oauth/userinfo", "GET", response_type="object")
-    if len(response) > 1 and 200 > response[1] > 300:
-        raise Exception(f"Received HTTP {response[1]}")
-    accounts = [
-        acct for acct in response[0]['accounts']
-        if acct['account_id'] == settings.DOCUSIGN_ACCOUNT_ID
-    ]
-    if not accounts:
-        return None
-    return accounts[0]['base_uri']
-
-
-def validate_and_set_consent_code(code: str) -> bool:
-    '''
-    Validate that the given consent code has the access permissions we need,
-    and set it as the current consent code for all our e-signing requests
-    with DocuSign.
-
-    Returns False if the consent code doesn't have the correct permissions.
-    '''
-
-    token = request_jwt_user_token(code)
-
-    base_uri = get_account_base_uri(token)
-    if not base_uri:
-        return False
-
-    cfg = get_config()
-    cfg.consent_code = code
-    cfg.consent_code_updated_at = timezone.now()
-    cfg.base_uri = base_uri
-    cfg.save()
-    return True
-
-
-def is_enabled() -> bool:
-    '''
-    Returns whether DocuSign integration is enabled.
-    '''
-
-    return bool(settings.DOCUSIGN_ACCOUNT_ID)
-
-
-def ensure_valid_configuration():
-    '''
-    Ensures that the DocuSign settings are properly defined. It doesn't actually
-    verify that *DocuSign* thinks they're valid, though.
-    '''
-
-    for setting in REQUIRED_SETTINGS:
-        if not getattr(settings, setting):
-            raise ImproperlyConfigured(f"The {setting} setting is not configured!")
-
-    config = get_config()
-
-    if not config.private_key:
-        raise ImproperlyConfigured(
-            "DocuSign private key is not configured! Please run the 'manage.py setdocusignkey' "
-            "command."
-        )
-
-    if not (config.consent_code and config.base_uri):
-        url = absolute_reverse('hpaction:docusign_consent')
-        raise ImproperlyConfigured(
-            f"DocuSign consent code is not configured!  "
-            f"Please obtain consent from a DocuSign user at {url}."
-        )
-
-
-def get_config() -> DocusignConfig:
-    '''
-    Return the singleton DocuSign configuration from the database.
-    '''
-
-    return DocusignConfig.objects.get()
-
-
-def get_auth_server_domain() -> str:
-    '''
-    Returns the domain name for the DocuSign authentication server.
-    '''
-
-    # TODO: We should return the production one if needed.
-    return 'account-d.docusign.com'
-
-
-def get_auth_server_url() -> str:
-    return f'https://{get_auth_server_domain()}'
-
-
-def get_private_key_bytes() -> bytes:
-    return get_config().private_key.encode('ascii')
-
-
-def docusign_client_user_id(user: JustfixUser) -> str:
-    '''
-    Given a user, return the DocuSign client user ID representing it.
-
-    According to the DocuSign documentation, the client user ID is:
-
-    > An alphanumeric value that serves as an ID for the signer throughout
-    > the signing process. The combination of this value, the signer's name,
-    > and the signer's email is used to uniquely identify them.
-
-    For more details, see:
-
-    https://developers.docusign.com/esign-rest-api/code-examples/quickstart-request-signature-embedded
-    '''
-
-    return str(user.pk)
-
-
-def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> docusign.EnvelopeDefinition:
+def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> dse.EnvelopeDefinition:
     '''
     Create a DocuSign envelope definition for the given HP Action documents.
     '''
@@ -162,14 +27,14 @@ def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> docusign.Enve
     pdf_bytes = docs.pdf_file.open().read()
     base64_pdf = base64.b64encode(pdf_bytes).decode('ascii')
 
-    document = docusign.Document(
+    document = dse.Document(
         document_base64=base64_pdf,
         name=f"HP Action forms for {user.full_name}",
         file_extension="pdf",
         document_id=HPA_DOCUMENT_ID,
     )
 
-    signer = docusign.Signer(
+    signer = dse.Signer(
         email=user.email,
         name=user.full_name,
         recipient_id=TENANT_RECIPIENT_ID,
@@ -181,7 +46,7 @@ def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> docusign.Enve
     # generated HP Action forms, creating fields using the drag-and-drop UI,
     # and noting their locations.
 
-    sign_here_petition = docusign.SignHere(
+    sign_here_petition = dse.SignHere(
         document_id=HPA_DOCUMENT_ID,
         page_number='4',
         recipient_id=TENANT_RECIPIENT_ID,
@@ -190,7 +55,7 @@ def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> docusign.Enve
         y_position='556',
     )
 
-    sign_here_verification = docusign.SignHere(
+    sign_here_verification = dse.SignHere(
         document_id=HPA_DOCUMENT_ID,
         page_number='4',
         recipient_id=TENANT_RECIPIENT_ID,
@@ -199,7 +64,7 @@ def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> docusign.Enve
         y_position='667',
     )
 
-    sign_here_hpd_inspection = docusign.SignHere(
+    sign_here_hpd_inspection = dse.SignHere(
         document_id=HPA_DOCUMENT_ID,
         page_number='5',
         recipient_id=TENANT_RECIPIENT_ID,
@@ -208,7 +73,7 @@ def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> docusign.Enve
         y_position='625',
     )
 
-    signer.tabs = docusign.Tabs(
+    signer.tabs = dse.Tabs(
         sign_here_tabs=[
             sign_here_petition,
             sign_here_verification,
@@ -216,57 +81,31 @@ def create_envelope_definition_for_hpa(docs: HPActionDocuments) -> docusign.Enve
         ],
     )
 
-    user_cc = docusign.CarbonCopy(
+    user_cc = dse.CarbonCopy(
         email=user.email,
         name=user.full_name,
         recipient_id="2",
         routing_order="2",
     )
 
-    envelope_definition = docusign.EnvelopeDefinition(
+    envelope_definition = dse.EnvelopeDefinition(
         email_subject=f"HP Action forms for {user.full_name}",
         documents=[document],
-        recipients=docusign.Recipients(signers=[signer], carbon_copies=[user_cc]),
+        recipients=dse.Recipients(signers=[signer], carbon_copies=[user_cc]),
         status="sent",
     )
 
-    assert isinstance(envelope_definition, docusign.EnvelopeDefinition)
+    assert isinstance(envelope_definition, dse.EnvelopeDefinition)
 
     return envelope_definition
 
 
-def create_api_client(
-    host: str,
-    access_token: Union[None, str, docusign.OAuthToken] = None
-) -> docusign.ApiClient:
-    api_client = docusign.ApiClient()
-    api_client.host = host
-    if isinstance(access_token, docusign.OAuthToken):
-        access_token = access_token.access_token
-    if access_token:
-        api_client.set_default_header('Authorization', f'Bearer {access_token}')
-    return api_client
-
-
-def create_default_api_client() -> docusign.ApiClient:
-    '''
-    Creates a DocuSign API client using the currently-configured settings.
-
-    This requests a JSON Web Token in the process, so it won't return
-    instantaneously.
-    '''
-
-    config = get_config()
-    token = request_jwt_user_token(config.consent_code)
-    return create_api_client(f"{config.base_uri}/restapi", token)
-
-
 def create_envelope_and_recipient_view_for_hpa(
     user: JustfixUser,
-    envelope_definition: docusign.EnvelopeDefinition,
-    api_client: docusign.ApiClient,
+    envelope_definition: dse.EnvelopeDefinition,
+    api_client: dse.ApiClient,
     return_url: str,
-) -> Tuple[docusign.EnvelopeSummary, str]:
+) -> Tuple[dse.EnvelopeSummary, str]:
     '''
     Create a DocuSign envelope and recipient view request for
     HP Action documents represented by a given envelope definition.
@@ -275,16 +114,16 @@ def create_envelope_and_recipient_view_for_hpa(
     URL to redirect the user to.
     '''
 
-    envelope_api = docusign.EnvelopesApi(api_client)
+    envelope_api = dse.EnvelopesApi(api_client)
     envelope = envelope_api.create_envelope(
         settings.DOCUSIGN_ACCOUNT_ID,
         envelope_definition=envelope_definition
     )
 
-    assert isinstance(envelope, docusign.EnvelopeSummary)
+    assert isinstance(envelope, dse.EnvelopeSummary)
 
     envelope_id = envelope.envelope_id
-    recipient_view_request = docusign.RecipientViewRequest(
+    recipient_view_request = dse.RecipientViewRequest(
         authentication_method='None',  # TODO: This should probably change?
         client_user_id=docusign_client_user_id(user),
         recipient_id=TENANT_RECIPIENT_ID,
@@ -300,44 +139,3 @@ def create_envelope_and_recipient_view_for_hpa(
     )
 
     return (envelope, results.url)
-
-
-def create_oauth_consent_url(
-    return_url: str,
-    state: str = '',
-) -> str:
-    '''
-    Returns a URL to redirect the user to, which will start the
-    DocuSign OAuth consent flow.
-    '''
-
-    qs = urllib.parse.urlencode({
-        'response_type': 'code',
-        'scope': 'signature impersonation',
-        'client_id': settings.DOCUSIGN_INTEGRATION_KEY,
-        'state': state,
-        'redirect_uri': return_url,
-    })
-    return f'{get_auth_server_url()}/oauth/auth?{qs}'
-
-
-def request_jwt_user_token(code: str) -> docusign.OAuthToken:
-    '''
-    Request a DocuSign JSON Web Token for the given user's consent code,
-    and return it.
-
-    Note tha this JWT will only last a short time (at the time of
-    this writing, it's 5 minutes). It's the consent code that's
-    long-lived, not the JWT.
-    '''
-
-    api_client = create_api_client(get_auth_server_url())
-    token = api_client.request_jwt_user_token(
-        client_id=settings.DOCUSIGN_INTEGRATION_KEY,
-        user_id=settings.DOCUSIGN_USER_ID,
-        oauth_host_name=get_auth_server_domain(),
-        private_key_bytes=get_private_key_bytes(),
-        expires_in=JWT_EXPIRATION,
-    )
-    assert isinstance(token, docusign.OAuthToken)
-    return token
